+++ plugins/woocommerce-admin/tests/js/setup-globals.js
@@ -44,7 +44,7 @@ wooCommercePackages.forEach( lib => {
 global.wcSettings = {
 	adminUrl: 'https://vagrant.local/wp/wp-admin/',
 	locale: 'en-US',
-	currency: { code: 'USD', precision: 2, symbol: '&#36;' },
+	currency: { code: 'USD', precision: 2, symbol: '$' },
 	date: {
 		dow: 0,
 	},