+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,3 +1,6 @@
+# 1.4.1 (unreleased)
+- Chart component: format numbers and prices using store currency settings.
+
 # 1.4.0
 - Add download log ip address autocompleter to search component
 - Add order number autocompleter to search component