+++ plugins/woocommerce-admin/packages/currency/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0 (unreleased)
+
+- Format using store currency settings (instead of locale)
+- Add optional currency symbol parameter
+
 # 1.0.0
 
 - Released package